+++ semcore/dropdown/CHANGELOG.md
@@ -8,6 +8,12 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - Version preminor update due to children dependencies update (`@semcore/utils` [4.8.4 ~> 4.9.0], `@semcore/core` [2.7.7 ~> 2.8.0]).
 
+## [4.8.4] - 2023-10-25
+
+### Changed
+
+- Prevent `Tab` keypress for opened dropdown with non-focusable content.
+
 ## [4.8.3] - 2023-10-24
 
 ### Changed