+++ semcore/time-picker/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.16.0] - 2023-10-31
+
+### Changed
+
+- Version preminor update due to children dependencies update (`@semcore/select` [4.15.0 ~> 4.16.0]).
+
 ## [4.15.0] - 2023-10-27
 
 ### Changed